+++ 0.56.b/graysvr/CBase.cpp
@@ -178,7 +178,7 @@ bool CBaseBaseDef::r_LoadVal( CScript & s )
 		case OBC_DAM:
 			{
 				int piVal[2];
-				int iQty = Str_ParseCmds( s.GetArgStr(), piVal, COUNTOF(piVal));
+				size_t iQty = Str_ParseCmds( s.GetArgStr(), piVal, COUNTOF(piVal));
 				m_attackBase = piVal[0];
 				if ( iQty > 1 )
 				{
@@ -210,7 +210,7 @@ bool CBaseBaseDef::r_LoadVal( CScript & s )
 		case OBC_RANGE:
 			{
 				int piVal[2];
-				int iQty = Str_ParseCmds( s.GetArgStr(), piVal, COUNTOF(piVal));
+				size_t iQty = Str_ParseCmds( s.GetArgStr(), piVal, COUNTOF(piVal));
 				if ( iQty > 1 )
 				{
 					m_range	 = ((piVal[0] & 0xff) << 8) & 0xff00;
